requirements.txt
requests==2.31.0
numpy==1.26.4
arabic_reshaper==3.0.0
python-bidi==0.4.2
PyMuPDF==1.23.25
google-generativeai==0.4.1
python-dotenv==1.0.1
fastapi==0.110.1
uvicorn==0.27.1
textblob==0.18.0
spacy==3.7.2
PyPDF2==3.0.1
